+++ src/Analyzer/ValidationUtils.cpp
@@ -1,0 +1,282 @@
+#include <Analyzer/ValidationUtils.h>
+
+#include <Analyzer/ConstantNode.h>
+#include <Analyzer/FunctionNode.h>
+#include <Analyzer/ColumnNode.h>
+#include <Analyzer/TableNode.h>
+#include <Analyzer/QueryNode.h>
+#include <Analyzer/InDepthQueryTreeVisitor.h>
+#include <Analyzer/AggregationUtils.h>
+#include <Analyzer/WindowFunctionsUtils.h>
+
+namespace DB
+{
+
+namespace ErrorCodes
+{
+    extern const int NOT_AN_AGGREGATE;
+    extern const int NOT_IMPLEMENTED;
+}
+
+class ValidateGroupByColumnsVisitor : public ConstInDepthQueryTreeVisitor<ValidateGroupByColumnsVisitor>
+{
+public:
+    explicit ValidateGroupByColumnsVisitor(const QueryTreeNodes & group_by_keys_nodes_, const QueryTreeNodePtr & query_node_)
+        : group_by_keys_nodes(group_by_keys_nodes_)
+        , query_node(query_node_)
+    {}
+
+    void visitImpl(const QueryTreeNodePtr & node)
+    {
+        auto query_tree_node_type = node->getNodeType();
+        if (query_tree_node_type == QueryTreeNodeType::CONSTANT ||
+            query_tree_node_type == QueryTreeNodeType::SORT ||
+            query_tree_node_type == QueryTreeNodeType::INTERPOLATE)
+            return;
+
+        if (nodeIsAggregateFunctionOrInGroupByKeys(node))
+            return;
+
+        auto * function_node = node->as<FunctionNode>();
+        if (function_node && function_node->getFunctionName() == "grouping")
+        {
+            auto & grouping_function_arguments_nodes = function_node->getArguments().getNodes();
+            for (auto & grouping_function_arguments_node : grouping_function_arguments_nodes)
+            {
+                bool found_argument_in_group_by_keys = false;
+
+                for (const auto & group_by_key_node : group_by_keys_nodes)
+                {
+                    if (grouping_function_arguments_node->isEqual(*group_by_key_node))
+                    {
+                        found_argument_in_group_by_keys = true;
+                        break;
+                    }
+                }
+
+                if (!found_argument_in_group_by_keys)
+                    throw Exception(ErrorCodes::NOT_AN_AGGREGATE,
+                        "GROUPING function argument {} is not in GROUP BY keys. In query {}",
+                        grouping_function_arguments_node->formatASTForErrorMessage(),
+                        query_node->formatASTForErrorMessage());
+            }
+
+            return;
+        }
+
+        auto * column_node = node->as<ColumnNode>();
+        if (!column_node)
+            return;
+
+        auto column_node_source = column_node->getColumnSource();
+        if (column_node_source->getNodeType() == QueryTreeNodeType::LAMBDA)
+            return;
+
+        throw Exception(ErrorCodes::NOT_AN_AGGREGATE,
+            "Column {} is not under aggregate function and not in GROUP BY keys. In query {}",
+            column_node->formatConvertedASTForErrorMessage(),
+            query_node->formatASTForErrorMessage());
+    }
+
+    bool needChildVisit(const QueryTreeNodePtr & parent_node, const QueryTreeNodePtr & child_node)
+    {
+        if (nodeIsAggregateFunctionOrInGroupByKeys(parent_node))
+            return false;
+
+        auto child_node_type = child_node->getNodeType();
+        return !(child_node_type == QueryTreeNodeType::QUERY || child_node_type == QueryTreeNodeType::UNION);
+    }
+
+private:
+    bool nodeIsAggregateFunctionOrInGroupByKeys(const QueryTreeNodePtr & node) const
+    {
+        if (auto * function_node = node->as<FunctionNode>())
+            if (function_node->isAggregateFunction())
+                return true;
+
+        for (const auto & group_by_key_node : group_by_keys_nodes)
+            if (node->isEqual(*group_by_key_node, {.compare_aliases = false}))
+                return true;
+
+        return false;
+    }
+
+    const QueryTreeNodes & group_by_keys_nodes;
+    const QueryTreeNodePtr & query_node;
+};
+
+void validateAggregates(const QueryTreeNodePtr & query_node)
+{
+    const auto & query_node_typed = query_node->as<QueryNode &>();
+    auto join_tree_node_type = query_node_typed.getJoinTree()->getNodeType();
+    bool join_tree_is_subquery = join_tree_node_type == QueryTreeNodeType::QUERY || join_tree_node_type == QueryTreeNodeType::UNION;
+
+    if (!join_tree_is_subquery)
+    {
+        assertNoAggregateFunctionNodes(query_node_typed.getJoinTree(), "in JOIN TREE");
+        assertNoGroupingFunctionNodes(query_node_typed.getJoinTree(), "in JOIN TREE");
+        assertNoWindowFunctionNodes(query_node_typed.getJoinTree(), "in JOIN TREE");
+    }
+
+    if (query_node_typed.hasWhere())
+    {
+        assertNoAggregateFunctionNodes(query_node_typed.getWhere(), "in WHERE");
+        assertNoGroupingFunctionNodes(query_node_typed.getWhere(), "in WHERE");
+        assertNoWindowFunctionNodes(query_node_typed.getWhere(), "in WHERE");
+    }
+
+    if (query_node_typed.hasPrewhere())
+    {
+        assertNoAggregateFunctionNodes(query_node_typed.getPrewhere(), "in PREWHERE");
+        assertNoGroupingFunctionNodes(query_node_typed.getPrewhere(), "in PREWHERE");
+        assertNoWindowFunctionNodes(query_node_typed.getPrewhere(), "in PREWHERE");
+    }
+
+    if (query_node_typed.hasHaving())
+        assertNoWindowFunctionNodes(query_node_typed.getHaving(), "in HAVING");
+
+    if (query_node_typed.hasWindow())
+        assertNoWindowFunctionNodes(query_node_typed.getWindowNode(), "in WINDOW");
+
+    QueryTreeNodes aggregate_function_nodes;
+    QueryTreeNodes window_function_nodes;
+
+    collectAggregateFunctionNodes(query_node, aggregate_function_nodes);
+    collectWindowFunctionNodes(query_node, window_function_nodes);
+
+    if (query_node_typed.hasGroupBy())
+    {
+        assertNoAggregateFunctionNodes(query_node_typed.getGroupByNode(), "in GROUP BY");
+        assertNoGroupingFunctionNodes(query_node_typed.getGroupByNode(), "in GROUP BY");
+        assertNoWindowFunctionNodes(query_node_typed.getGroupByNode(), "in GROUP BY");
+    }
+
+    for (auto & aggregate_function_node : aggregate_function_nodes)
+    {
+        auto & aggregate_function_node_typed = aggregate_function_node->as<FunctionNode &>();
+
+        assertNoAggregateFunctionNodes(aggregate_function_node_typed.getArgumentsNode(), "inside another aggregate function");
+        assertNoGroupingFunctionNodes(aggregate_function_node_typed.getArgumentsNode(), "inside another aggregate function");
+        assertNoWindowFunctionNodes(aggregate_function_node_typed.getArgumentsNode(), "inside an aggregate function");
+    }
+
+    for (auto & window_function_node : window_function_nodes)
+    {
+        auto & window_function_node_typed = window_function_node->as<FunctionNode &>();
+        assertNoWindowFunctionNodes(window_function_node_typed.getArgumentsNode(), "inside another window function");
+
+        if (query_node_typed.hasWindow())
+            assertNoWindowFunctionNodes(window_function_node_typed.getWindowNode(), "inside window definition");
+    }
+
+    QueryTreeNodes group_by_keys_nodes;
+    group_by_keys_nodes.reserve(query_node_typed.getGroupBy().getNodes().size());
+
+    for (const auto & node : query_node_typed.getGroupBy().getNodes())
+    {
+        if (query_node_typed.isGroupByWithGroupingSets())
+        {
+            auto & grouping_set_keys = node->as<ListNode &>();
+            for (auto & grouping_set_key : grouping_set_keys.getNodes())
+            {
+                if (grouping_set_key->as<ConstantNode>())
+                    continue;
+
+                group_by_keys_nodes.push_back(grouping_set_key);
+            }
+        }
+        else
+        {
+            if (node->as<ConstantNode>())
+                continue;
+
+            group_by_keys_nodes.push_back(node);
+        }
+    }
+
+    if (query_node_typed.getGroupBy().getNodes().empty())
+    {
+        if (query_node_typed.hasHaving())
+            assertNoGroupingFunctionNodes(query_node_typed.getHaving(), "in HAVING without GROUP BY");
+
+        if (query_node_typed.hasOrderBy())
+            assertNoGroupingFunctionNodes(query_node_typed.getOrderByNode(), "in ORDER BY without GROUP BY");
+
+        assertNoGroupingFunctionNodes(query_node_typed.getProjectionNode(), "in SELECT without GROUP BY");
+    }
+
+    bool has_aggregation = !query_node_typed.getGroupBy().getNodes().empty() || !aggregate_function_nodes.empty();
+
+    if (has_aggregation)
+    {
+        ValidateGroupByColumnsVisitor validate_group_by_columns_visitor(group_by_keys_nodes, query_node);
+
+        if (query_node_typed.hasHaving())
+            validate_group_by_columns_visitor.visit(query_node_typed.getHaving());
+
+        if (query_node_typed.hasOrderBy())
+            validate_group_by_columns_visitor.visit(query_node_typed.getOrderByNode());
+
+        validate_group_by_columns_visitor.visit(query_node_typed.getProjectionNode());
+    }
+
+    bool aggregation_with_rollup_or_cube_or_grouping_sets = query_node_typed.isGroupByWithRollup() ||
+        query_node_typed.isGroupByWithCube() ||
+        query_node_typed.isGroupByWithGroupingSets();
+    if (!has_aggregation && (query_node_typed.isGroupByWithTotals() || aggregation_with_rollup_or_cube_or_grouping_sets))
+        throw Exception(ErrorCodes::NOT_IMPLEMENTED, "WITH TOTALS, ROLLUP, CUBE or GROUPING SETS are not supported without aggregation");
+}
+
+namespace
+{
+
+class ValidateFunctionNodesVisitor : public ConstInDepthQueryTreeVisitor<ValidateFunctionNodesVisitor>
+{
+public:
+    explicit ValidateFunctionNodesVisitor(std::string_view function_name_,
+        int exception_code_,
+        std::string_view exception_function_name_,
+        std::string_view exception_place_message_)
+        : function_name(function_name_)
+        , exception_code(exception_code_)
+        , exception_function_name(exception_function_name_)
+        , exception_place_message(exception_place_message_)
+    {}
+
+    void visitImpl(const QueryTreeNodePtr & node)
+    {
+        auto * function_node = node->as<FunctionNode>();
+        if (function_node && function_node->getFunctionName() == function_name)
+            throw Exception(exception_code,
+                "{} function {} is found {} in query",
+                exception_function_name,
+                function_node->formatASTForErrorMessage(),
+                exception_place_message);
+    }
+
+    static bool needChildVisit(const QueryTreeNodePtr &, const QueryTreeNodePtr & child_node)
+    {
+        auto child_node_type = child_node->getNodeType();
+        return !(child_node_type == QueryTreeNodeType::QUERY || child_node_type == QueryTreeNodeType::UNION);
+    }
+
+private:
+    std::string_view function_name;
+    int exception_code = 0;
+    std::string_view exception_function_name;
+    std::string_view exception_place_message;
+};
+
+}
+
+void assertNoFunctionNodes(const QueryTreeNodePtr & node,
+    std::string_view function_name,
+    int exception_code,
+    std::string_view exception_function_name,
+    std::string_view exception_place_message)
+{
+    ValidateFunctionNodesVisitor visitor(function_name, exception_code, exception_function_name, exception_place_message);
+    visitor.visit(node);
+}
+
+}
