+++ tests/queries/0_stateless/02479_analyzer_join_with_constants.reference
@@ -8,8 +8,3 @@
 --
 1	1	0	0
 --
-0
-0
-0
-0
-0
